Git commit for the first major development of the Biofuel Engine Control System software.
The code has been split into two libraries:
- control_system_methods.h, which implements basic functions for measuring sensors and opening/closing circuits. It also defines a new type called 'engine', which holds vital information about the engine's state, such as its speed, position and the pins associated with the coils and injectors.

- cbr600f4i_operating_map.h, which implements a new typedef called 'operating_point', which contains the spark/fuel timings for a particular target RPM. In the library, an array of operating points has been defined called the operating_map, which is optimised for the CBR600f4i to run on ethanol. Furthermore the library contains functions for getting the correct timings based on a target RPM, as well as opening/closing the approriate circuit based on the crankshaft angle.

More work needs to be done to define some of the functions declared in the header files. Timer interrupts need to be implemented for the fuel duration and dwell time and a new way of cross-checking the CPG signals against the current crank angle needs to be designed, as the current version is not fool-proof against missed IPG pulses.

Codes for the scripts will be implemented once this version of the software is completed.